% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fit-survival.R
\docType{package}
\name{triageR-package}
\alias{triageR}
\alias{triageR-package}
\title{Fit a clinical survival (time-to-event) model}
\description{
triageR: Survival Analysis Utilities
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DevWebWacky/triageR}
  \item Report bugs at \url{https://github.com/DevWebWacky/triageR/issues}
}

}
\author{
\strong{Maintainer}: Uwakmfon Paul \email{uwakmfon31@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Uwakmfon Paul \email{uwakmfon31@gmail.com} [copyright holder]
}

}
\keyword{internal}
